Refactor RSI into class
